+++ lib/generators/auxiliary_rails/install_generator.rb
@@ -5,6 +5,7 @@ module AuxiliaryRails
     def install
       generate 'auxiliary_rails:install_commands'
       generate 'auxiliary_rails:install_errors'
+      generate 'auxiliary_rails:install_errors_controller'
     end
   end
 end