+++ lib/generators/auxiliary_rails/templates/errors_controller/errors_controller_template.rb
@@ -1,0 +1,15 @@
+class ErrorsController < ApplicationController
+  DEFAULT_ERROR_STATUS = 500
+
+  def error
+    render status: params[:status] || DEFAULT_ERROR_STATUS
+  end
+
+  def not_found_error
+    render status: 404
+  end
+
+  def unacceptable_error
+    render status: 422
+  end
+end